Add summaries for the simple methods
