+++ src/SimdLinq/SimdLinqExtensions.LongSum.tt
@@ -1,0 +1,44 @@
+﻿<#@ template debug="false" hostspecific="false" language="C#" #>
+<#@ assembly name="System.Core" #>
+<#@ import namespace="System.Linq" #>
+<#@ import namespace="System.Text" #>
+<#@ import namespace="System.Collections.Generic" #>
+<#@ output extension=".cs" #>
+<#
+    string summary(string x, bool nullable)
+    {
+        var exception = "";
+        if (nullable)
+        {
+            exception += """
+
+               /// <exception cref="ArgumentNullException"><paramref name="source"/> is null.</exception>
+           """;
+        }
+
+        return $"""
+        /// <summary>
+            /// Computes the long sum of a collection of <see cref="{x}"/> values using SIMD acceleration.
+            /// </summary>
+            /// <remarks>
+            /// <para>Note that unlike Linq.Sum this method is unchecked and will not throw an <see cref="OverflowException"/>.</para>
+            /// </remarks>
+            /// <param name="source">A collection of <see cref="{x}"/> values to calculate the long sum of.</param>
+            /// <returns>The sum of the values in the collection.</returns>{exception}
+        """;
+    }
+
+    var retType = (string x) => x is "int" ? "long" : "ulong";;
+   
+    var methodName = "LongSum";
+    var types = new []
+    {
+        "int", "uint"
+    };
+#>
+namespace SimdLinq;
+
+using System.Runtime.InteropServices;
+using System.Runtime.Intrinsics;
+
+<#@ include file=".\SimdLinqExtensionsSkele.ttinclude" #>
