--- src/SimdLinq/SimdLinqExtensions.SumAverage.tt
@@ -1,28 +1,0 @@
-﻿<#@ template debug="false" hostspecific="false" language="C#" #>
-<#@ assembly name="System.Core" #>
-<#@ import namespace="System.Linq" #>
-<#@ import namespace="System.Text" #>
-<#@ import namespace="System.Collections.Generic" #>
-<#@ output extension=".cs" #>
-<#
-    string methodName = "Sum";
-    var retType = (string x) => x;
-    var types = new []
-    {
-        "int", "long", "double", "float", "uint", "ulong"
-    };
-#>
-namespace SimdLinq;
-
-using System.Runtime.InteropServices;
-
-<#@ include file=".\SimdLinqExtensions.ttinclude" #>
-
-<# methodName = "Average"; #>
-<# retType = (string x) => x is "double" or "float" ? x : "double"; #>
-<#@ include file=".\SimdLinqExtensions.ttinclude" #>
-
-<# methodName = "LongSum"; #>
-<# retType = (string x) => x is "int" ? "long" : "ulong"; #>
-<# types = new [] { "int", "uint" }; #>
-<#@ include file=".\SimdLinqExtensions.ttinclude" #>
